New Drive Train Programming
Got rid of arcade drive for driving the robot. Instead, a custom drive
class was made, using lots of if/else if statments to control the
motors. This allows for simple, yet complex changing of sensitivity on
the controller for driving. More changes were also made to the claw
motor programming, to make it work right.